Added explanation ability
I added the ability to give explanations for the bot's actions. I had to make some changes to make the explanation box visible, as the settings box hid it (these changes are to do with the range sliders, as I had to reduce their height to show the box). Since the settings box is now smaller, the tooltips covered the setting names, so I made them invisible until a user hovers over their sliders